# api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LongLink Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/applications": {
      "get": {
        "summary": "List Applications",
        "description": "Return all applications for administrator views.",
        "operationId": "list_applications_api_v1_applications_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ApplicationResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/applications": {
      "post": {
        "summary": "Create Application",
        "description": "Create Application state and queue its explicit deployment lifecycle.",
        "operationId": "create_application_api_v1_organizations__organization_id__applications_post",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Organization Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Organization Applications",
        "description": "Return applications visible to the current organization member.",
        "operationId": "get_organization_applications_api_v1_organizations__organization_id__applications_get",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Organization Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationApplicationSummary"
                  },
                  "title": "Response Get Organization Applications Api V1 Organizations  Organization Id  Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/releases": {
      "post": {
        "summary": "Release Application",
        "description": "Record one desired Application release and queue its deployment.",
        "operationId": "release_application_api_v1_applications__application_id__releases_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationRelease"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/logs": {
      "get": {
        "summary": "Get Application Logs",
        "description": "Return recent pod logs for one managed application.",
        "operationId": "get_application_logs_api_v1_applications__application_id__logs_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Application Logs Api V1 Applications  Application Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}": {
      "delete": {
        "summary": "Delete Application",
        "description": "Mark one Application absent and queue explicit lifecycle cleanup.",
        "operationId": "delete_application_api_v1_applications__application_id__delete",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "icon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Icon"
              },
              {
                "type": "null"
              }
            ]
          },
          "image": {
            "type": "string",
            "title": "Image"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "envs": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Envs"
          }
        },
        "type": "object",
        "required": [
          "name",
          "image"
        ],
        "title": "ApplicationCreate",
        "description": "Validate application creation payloads."
      },
      "ApplicationRelease": {
        "properties": {
          "image": {
            "type": "string",
            "title": "Image"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "ApplicationRelease",
        "description": "Validate a requested Application release."
      },
      "ApplicationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationIdentity"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "icon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Icon"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "image_desired": {
            "type": "string",
            "title": "Image Desired"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization",
          "name",
          "slug",
          "icon",
          "description",
          "image_desired",
          "status",
          "created_at"
        ],
        "title": "ApplicationResponse",
        "description": "Represent one application in API responses."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Icon": {
        "type": "string",
        "enum": [
          "activity",
          "arrow-right",
          "banknote",
          "bell",
          "box",
          "boxes",
          "building-2",
          "check",
          "clipboard-list",
          "container",
          "cpu",
          "database",
          "download",
          "hard-drive",
          "layers",
          "layout-dashboard",
          "layout-grid",
          "link",
          "list",
          "list-check",
          "map-pin",
          "plus",
          "rocket",
          "rotate-ccw",
          "settings-2",
          "shield-check",
          "sliders-horizontal",
          "timer",
          "users",
          "x"
        ],
        "title": "Icon",
        "description": "Lucide icon slugs supported by LongLink runtimes."
      },
...
      "OrganizationApplicationSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "icon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Icon"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "status"
        ],
        "title": "OrganizationApplicationSummary",
        "description": "Represent a compact LongLink Application in nested Organization responses."
      },
...
      "OrganizationIdentity": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "avatar": {
            "type": "string",
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "avatar"
        ],
        "title": "OrganizationIdentity",
        "description": "Represent a compact Organization in nested API responses."
      },
...
      "Page_ApplicationResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[ApplicationResponse]"
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "creating",
          "running"
        ],
        "title": "Status",
        "description": "Lifecycle states shared by Platform-managed resources."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}